Move application logic from Web to Core
- Rename Parser project to Core.
- Move application logic.
