+++ Money.Core/Handlers/SaveStatementHandler.cs
@@ -1,8 +1,7 @@
 ﻿using MediatR;
-using Money.Db;
-using Money.Web.Requests;
+using Money.Core.Requests;
 
-namespace Money.Web.Handlers
+namespace Money.Core.Handlers
 {
   public class SaveStatementHandler : IRequestHandler<SaveExpensesRequest>
   {
