+++ Money.Core/Money.Core.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{948C0B9B-150A-4470-9145-16EBE47C54E4}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Money</RootNamespace>
-    <AssemblyName>Money</AssemblyName>
+    <RootNamespace>Money.Core</RootNamespace>
+    <AssemblyName>Money.Core</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -36,6 +36,9 @@
     <Reference Include="itextsharp, Version=5.5.12.0, Culture=neutral, PublicKeyToken=8354ae6d2174ddca, processorArchitecture=MSIL">
       <HintPath>..\packages\iTextSharp.5.5.12\lib\itextsharp.dll</HintPath>
     </Reference>
+    <Reference Include="MediatR, Version=3.0.1.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\MediatR.3.0.1\lib\net45\MediatR.dll</HintPath>
+    </Reference>
     <Reference Include="MongoDB.Bson, Version=2.5.0.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\MongoDB.Bson.2.5.0\lib\net45\MongoDB.Bson.dll</HintPath>
     </Reference>
@@ -57,13 +60,22 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IStatementService.cs" />
+    <Compile Include="Handlers\DeleteExpenseHandler.cs" />
+    <Compile Include="Handlers\GetExpensesHandler.cs" />
+    <Compile Include="Handlers\ParsePdfHandler.cs" />
+    <Compile Include="Handlers\SaveStatementHandler.cs" />
+    <Compile Include="IDbService.cs" />
+    <Compile Include="Services\IStatementService.cs" />
     <Compile Include="Models\Statement.cs" />
-    <Compile Include="StatementParsers\CirclekStatementParser.cs" />
-    <Compile Include="StatementParsers\IStatementParser.cs" />
-    <Compile Include="StatementParsers\SkandiaStatementParser.cs" />
+    <Compile Include="Notifications\ExpensesNotification.cs" />
+    <Compile Include="Requests\GetExpensesRequest.cs" />
+    <Compile Include="Requests\ParsePdfRequest.cs" />
+    <Compile Include="Requests\SaveExpensesRequest.cs" />
+    <Compile Include="Services\CirclekStatementParser.cs" />
+    <Compile Include="Services\IStatementParser.cs" />
+    <Compile Include="Services\SkandiaStatementParser.cs" />
     <Compile Include="Models\Expense.cs" />
-    <Compile Include="StatementService.cs" />
+    <Compile Include="Services\StatementService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
