+++ Money.Core/packages.config
@@ -2,6 +2,7 @@
 <packages>
   <package id="DnsClient" version="1.0.7" targetFramework="net461" />
   <package id="iTextSharp" version="5.5.12" targetFramework="net461" />
+  <package id="MediatR" version="3.0.1" targetFramework="net461" />
   <package id="MongoDB.Bson" version="2.5.0" targetFramework="net461" />
   <package id="MongoDB.Driver" version="2.5.0" targetFramework="net461" />
   <package id="MongoDB.Driver.Core" version="2.5.0" targetFramework="net461" />
