+++ Money.Db/Money.Db.csproj
@@ -75,7 +75,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IDbService.cs" />
     <Compile Include="MongoDbService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -92,9 +91,9 @@
     </BootstrapperPackage>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Money.Parser\Money.Parser.csproj">
+    <ProjectReference Include="..\Money.Core\Money.Core.csproj">
       <Project>{948c0b9b-150a-4470-9145-16ebe47c54e4}</Project>
-      <Name>Money.Parser</Name>
+      <Name>Money.Core</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
