+++ Money.Web/Money.Web.csproj
@@ -63,15 +63,6 @@
     <Reference Include="Microsoft.Owin.Security, Version=3.1.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\packages\Microsoft.Owin.Security.3.1.0\lib\net45\Microsoft.Owin.Security.dll</HintPath>
     </Reference>
-    <Reference Include="MongoDB.Bson, Version=2.5.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\MongoDB.Bson.2.5.0\lib\net45\MongoDB.Bson.dll</HintPath>
-    </Reference>
-    <Reference Include="MongoDB.Driver, Version=2.5.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\MongoDB.Driver.2.5.0\lib\net45\MongoDB.Driver.dll</HintPath>
-    </Reference>
-    <Reference Include="MongoDB.Driver.Core, Version=2.5.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\MongoDB.Driver.Core.2.5.0\lib\net45\MongoDB.Driver.Core.dll</HintPath>
-    </Reference>
     <Reference Include="Nancy, Version=1.4.4.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\Nancy.1.4.4\lib\net40\Nancy.dll</HintPath>
     </Reference>
@@ -151,28 +142,20 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Bootstrapper.cs" />
-    <Compile Include="Handlers\GetExpensesHandler.cs" />
-    <Compile Include="Handlers\ParsePdfHandler.cs" />
-    <Compile Include="Handlers\DeleteExpenseHandler.cs" />
-    <Compile Include="Handlers\SaveStatementHandler.cs" />
     <Compile Include="Modules\ExpensesModule.cs" />
-    <Compile Include="Notifications\ExpensesNotification.cs" />
-    <Compile Include="Requests\GetExpensesRequest.cs" />
     <Compile Include="Modules\ContentModule.cs" />
     <Compile Include="Modules\UploadModule.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Requests\ParsePdfRequest.cs" />
-    <Compile Include="Requests\SaveExpensesRequest.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\Money.Core\Money.Core.csproj">
+      <Project>{948c0b9b-150a-4470-9145-16ebe47c54e4}</Project>
+      <Name>Money.Core</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Money.Db\Money.Db.csproj">
       <Project>{09A55C14-A71D-4C0A-9834-526B835B91EE}</Project>
       <Name>Money.Db</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Money.Parser\Money.Parser.csproj">
-      <Project>{948c0b9b-150a-4470-9145-16ebe47c54e4}</Project>
-      <Name>Money</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup />
