+++ Money.Web/packages.config
@@ -8,9 +8,6 @@
   <package id="Microsoft.Owin" version="3.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="3.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security" version="3.1.0" targetFramework="net461" />
-  <package id="MongoDB.Bson" version="2.5.0" targetFramework="net461" />
-  <package id="MongoDB.Driver" version="2.5.0" targetFramework="net461" />
-  <package id="MongoDB.Driver.Core" version="2.5.0" targetFramework="net461" />
   <package id="Nancy" version="1.4.4" targetFramework="net461" />
   <package id="Nancy.Bootstrappers.StructureMap" version="1.4.3" targetFramework="net461" />
   <package id="Nancy.Hosting.Aspnet" version="1.4.1" targetFramework="net461" />
